+++ SisFactArrocan_LBD/src/main/java/Interfaz/Menu.java
@@ -1,0 +1,566 @@
+/*
+ * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
+ * Click nbfs://nbhost/SystemFileSystem/Templates/GUIForms/JFrame.java to edit this template
+ */
+package Interfaz;
+
+import java.awt.Color;
+
+/**
+ *
+ * @author gblan
+ */
+public class Menu extends javax.swing.JFrame {
+
+    /**
+     * Creates new form Menu
+     */
+    int xPos, yPos;
+
+    public Menu() {
+        initComponents();
+    }
+
+    /**
+     * This method is called from within the constructor to initialize the form.
+     * WARNING: Do NOT modify this code. The content of this method is always
+     * regenerated by the Form Editor.
+     */
+    @SuppressWarnings("unchecked")
+    // <editor-fold defaultstate="collapsed" desc="Generated Code">//GEN-BEGIN:initComponents
+    private void initComponents() {
+
+        bg = new javax.swing.JPanel();
+        header = new javax.swing.JPanel();
+        logo = new javax.swing.JLabel();
+        logoutText = new javax.swing.JLabel();
+        userText = new javax.swing.JLabel();
+        controlBar = new javax.swing.JPanel();
+        exitButton = new javax.swing.JPanel();
+        exitText = new javax.swing.JLabel();
+        tittle = new javax.swing.JLabel();
+        footer = new javax.swing.JPanel();
+        copyRight = new javax.swing.JLabel();
+        cardCliente = new Interfaz.PanelRound();
+        imgClienteBtn = new Interfaz.PanelRound();
+        imgClienteText = new javax.swing.JLabel();
+        clienteBtn = new Interfaz.PanelRound();
+        clienteText = new javax.swing.JLabel();
+        cardInventario = new Interfaz.PanelRound();
+        imgInventarioBtn = new Interfaz.PanelRound();
+        imgClienteText1 = new javax.swing.JLabel();
+        inventarioBtn = new Interfaz.PanelRound();
+        inventarioText = new javax.swing.JLabel();
+        cardFactura = new Interfaz.PanelRound();
+        imgFacturaBtn = new Interfaz.PanelRound();
+        imgFacturaText = new javax.swing.JLabel();
+        facturaBtn = new Interfaz.PanelRound();
+        facturaText = new javax.swing.JLabel();
+
+        setDefaultCloseOperation(javax.swing.WindowConstants.EXIT_ON_CLOSE);
+        setUndecorated(true);
+        setPreferredSize(new java.awt.Dimension(1280, 720));
+
+        bg.setBackground(new java.awt.Color(16, 36, 61));
+        bg.setLayout(new org.netbeans.lib.awtextra.AbsoluteLayout());
+
+        header.setBackground(new java.awt.Color(217, 90, 83));
+
+        logo.setIcon(new javax.swing.JLabel() {
+            public javax.swing.Icon getIcon() {
+                try {
+                    return new javax.swing.ImageIcon(
+                        new java.net.URL("https://blogger.googleusercontent.com/img/a/AVvXsEhn6YsZ5kg9LG29At7L51WfXN8DcypDDfWc3ba-IWw2W1BRnSWkv_7voMZmwHjjvY3KGDtFW4JkzCBLglVKnjfpHF7Aaz5VVKE3UgnLJBiliG7w_AWGMjHPMwqv1wcX3u_NnlzKzmx6lL8znw2Ne7gU9GSmv7Y5au4nGbrCxQKRafZpPwLSQ7SSujfiD7ut=w203-h100")
+                    );
+                } catch (java.net.MalformedURLException e) {
+                }
+                return null;
+            }
+        }.getIcon());
+
+        logoutText.setFont(new java.awt.Font("Segoe UI", 1, 24)); // NOI18N
+        logoutText.setForeground(new java.awt.Color(255, 255, 255));
+        logoutText.setText("Cerrar Sesión");
+
+        userText.setFont(new java.awt.Font("Segoe UI", 1, 24)); // NOI18N
+        userText.setForeground(new java.awt.Color(255, 255, 255));
+        userText.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
+        userText.setText("Gerald Blanco");
+
+        javax.swing.GroupLayout headerLayout = new javax.swing.GroupLayout(header);
+        header.setLayout(headerLayout);
+        headerLayout.setHorizontalGroup(
+            headerLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(headerLayout.createSequentialGroup()
+                .addComponent(logo, javax.swing.GroupLayout.PREFERRED_SIZE, 195, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, 740, Short.MAX_VALUE)
+                .addComponent(logoutText)
+                .addGap(26, 26, 26)
+                .addComponent(userText, javax.swing.GroupLayout.PREFERRED_SIZE, 161, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addContainerGap())
+        );
+        headerLayout.setVerticalGroup(
+            headerLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addComponent(logo, javax.swing.GroupLayout.DEFAULT_SIZE, 110, Short.MAX_VALUE)
+            .addComponent(logoutText, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+            .addComponent(userText, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+        );
+
+        bg.add(header, new org.netbeans.lib.awtextra.AbsoluteConstraints(0, 30, 1280, 110));
+
+        controlBar.setBackground(new java.awt.Color(217, 90, 83));
+        controlBar.addMouseMotionListener(new java.awt.event.MouseMotionAdapter() {
+            public void mouseDragged(java.awt.event.MouseEvent evt) {
+                controlBarMouseDragged(evt);
+            }
+        });
+        controlBar.addMouseListener(new java.awt.event.MouseAdapter() {
+            public void mousePressed(java.awt.event.MouseEvent evt) {
+                controlBarMousePressed(evt);
+            }
+        });
+
+        exitButton.setBackground(new java.awt.Color(217, 90, 83));
+
+        exitText.setFont(new java.awt.Font("Segoe UI", 1, 18)); // NOI18N
+        exitText.setForeground(new java.awt.Color(255, 255, 255));
+        exitText.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
+        exitText.setText("X");
+        exitText.addMouseListener(new java.awt.event.MouseAdapter() {
+            public void mouseClicked(java.awt.event.MouseEvent evt) {
+                exitTextMouseClicked(evt);
+            }
+            public void mouseEntered(java.awt.event.MouseEvent evt) {
+                exitTextMouseEntered(evt);
+            }
+            public void mouseExited(java.awt.event.MouseEvent evt) {
+                exitTextMouseExited(evt);
+            }
+        });
+
+        javax.swing.GroupLayout exitButtonLayout = new javax.swing.GroupLayout(exitButton);
+        exitButton.setLayout(exitButtonLayout);
+        exitButtonLayout.setHorizontalGroup(
+            exitButtonLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addComponent(exitText, javax.swing.GroupLayout.DEFAULT_SIZE, 40, Short.MAX_VALUE)
+        );
+        exitButtonLayout.setVerticalGroup(
+            exitButtonLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addComponent(exitText, javax.swing.GroupLayout.DEFAULT_SIZE, 30, Short.MAX_VALUE)
+        );
+
+        javax.swing.GroupLayout controlBarLayout = new javax.swing.GroupLayout(controlBar);
+        controlBar.setLayout(controlBarLayout);
+        controlBarLayout.setHorizontalGroup(
+            controlBarLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(controlBarLayout.createSequentialGroup()
+                .addComponent(exitButton, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addGap(0, 1240, Short.MAX_VALUE))
+        );
+        controlBarLayout.setVerticalGroup(
+            controlBarLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, controlBarLayout.createSequentialGroup()
+                .addGap(0, 0, Short.MAX_VALUE)
+                .addComponent(exitButton, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
+        );
+
+        bg.add(controlBar, new org.netbeans.lib.awtextra.AbsoluteConstraints(0, 0, 1280, 30));
+
+        tittle.setFont(new java.awt.Font("Segoe UI", 1, 36)); // NOI18N
+        tittle.setForeground(new java.awt.Color(255, 255, 255));
+        tittle.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
+        tittle.setText("Servicios");
+        bg.add(tittle, new org.netbeans.lib.awtextra.AbsoluteConstraints(530, 160, 210, 70));
+
+        footer.setBackground(new java.awt.Color(217, 90, 83));
+
+        copyRight.setFont(new java.awt.Font("Segoe UI", 1, 18)); // NOI18N
+        copyRight.setForeground(new java.awt.Color(255, 255, 255));
+        copyRight.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
+        copyRight.setText("Todos los derechos reservados 2023©");
+
+        javax.swing.GroupLayout footerLayout = new javax.swing.GroupLayout(footer);
+        footer.setLayout(footerLayout);
+        footerLayout.setHorizontalGroup(
+            footerLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(footerLayout.createSequentialGroup()
+                .addGap(385, 385, 385)
+                .addComponent(copyRight, javax.swing.GroupLayout.PREFERRED_SIZE, 490, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addContainerGap(405, Short.MAX_VALUE))
+        );
+        footerLayout.setVerticalGroup(
+            footerLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(footerLayout.createSequentialGroup()
+                .addComponent(copyRight, javax.swing.GroupLayout.DEFAULT_SIZE, 44, Short.MAX_VALUE)
+                .addContainerGap())
+        );
+
+        bg.add(footer, new org.netbeans.lib.awtextra.AbsoluteConstraints(0, 680, 1280, 50));
+
+        cardCliente.setBackground(new java.awt.Color(255, 255, 255));
+        cardCliente.setForeground(new java.awt.Color(255, 255, 255));
+        cardCliente.setRoundBottomLeft(20);
+        cardCliente.setRoundBottomRight(20);
+        cardCliente.setRoundTopLeft(20);
+        cardCliente.setRoundTopRight(20);
+
+        imgClienteBtn.setBackground(new java.awt.Color(255, 255, 255));
+        imgClienteBtn.setRoundTopLeft(20);
+        imgClienteBtn.setRoundTopRight(20);
+
+        imgClienteText.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
+        imgClienteText.setIcon(new javax.swing.JLabel() {
+            public javax.swing.Icon getIcon() {
+                try {
+                    return new javax.swing.ImageIcon(
+                        new java.net.URL("https://blogger.googleusercontent.com/img/a/AVvXsEiwMTkKkpAxKJun-cK_Fv7iV-5Mzm5tx4Heh7SRLWC8XwtuDNatMlnYS2qprAb43tNXbcmAQBzevp5u2ZTR6my5GPEDPaxNxB1Xo3a0ylrZD5AHsm7Ll8hLkAK8cVOkUUcZ03AGfRoSNabQXoZSinR2ImPKZxU6jql1Wm0BlN4dvc7BOyFLvW2siD81CmGq=w268-h234")
+                    );
+                } catch (java.net.MalformedURLException e) {
+                }
+                return null;
+            }
+        }.getIcon());
+
+        javax.swing.GroupLayout imgClienteBtnLayout = new javax.swing.GroupLayout(imgClienteBtn);
+        imgClienteBtn.setLayout(imgClienteBtnLayout);
+        imgClienteBtnLayout.setHorizontalGroup(
+            imgClienteBtnLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, imgClienteBtnLayout.createSequentialGroup()
+                .addContainerGap()
+                .addComponent(imgClienteText, javax.swing.GroupLayout.PREFERRED_SIZE, 0, Short.MAX_VALUE)
+                .addContainerGap())
+        );
+        imgClienteBtnLayout.setVerticalGroup(
+            imgClienteBtnLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(imgClienteBtnLayout.createSequentialGroup()
+                .addContainerGap()
+                .addComponent(imgClienteText, javax.swing.GroupLayout.DEFAULT_SIZE, 168, Short.MAX_VALUE))
+        );
+
+        clienteBtn.setBackground(new java.awt.Color(255, 255, 255));
+        clienteBtn.setRoundBottomLeft(20);
+        clienteBtn.setRoundBottomRight(20);
+
+        clienteText.setFont(new java.awt.Font("Segoe UI", 1, 24)); // NOI18N
+        clienteText.setForeground(new java.awt.Color(0, 0, 0));
+        clienteText.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
+        clienteText.setText("Clientes");
+
+        javax.swing.GroupLayout clienteBtnLayout = new javax.swing.GroupLayout(clienteBtn);
+        clienteBtn.setLayout(clienteBtnLayout);
+        clienteBtnLayout.setHorizontalGroup(
+            clienteBtnLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addComponent(clienteText, javax.swing.GroupLayout.DEFAULT_SIZE, 230, Short.MAX_VALUE)
+        );
+        clienteBtnLayout.setVerticalGroup(
+            clienteBtnLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, clienteBtnLayout.createSequentialGroup()
+                .addContainerGap()
+                .addComponent(clienteText, javax.swing.GroupLayout.DEFAULT_SIZE, 88, Short.MAX_VALUE)
+                .addContainerGap())
+        );
+
+        javax.swing.GroupLayout cardClienteLayout = new javax.swing.GroupLayout(cardCliente);
+        cardCliente.setLayout(cardClienteLayout);
+        cardClienteLayout.setHorizontalGroup(
+            cardClienteLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addComponent(imgClienteBtn, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+            .addComponent(clienteBtn, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+        );
+        cardClienteLayout.setVerticalGroup(
+            cardClienteLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(cardClienteLayout.createSequentialGroup()
+                .addComponent(imgClienteBtn, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addComponent(clienteBtn, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
+        );
+
+        bg.add(cardCliente, new org.netbeans.lib.awtextra.AbsoluteConstraints(120, 270, 230, 280));
+
+        cardInventario.setBackground(new java.awt.Color(255, 255, 255));
+        cardInventario.setForeground(new java.awt.Color(255, 255, 255));
+        cardInventario.setRoundBottomLeft(20);
+        cardInventario.setRoundBottomRight(20);
+        cardInventario.setRoundTopLeft(20);
+        cardInventario.setRoundTopRight(20);
+
+        imgInventarioBtn.setBackground(new java.awt.Color(255, 255, 255));
+        imgInventarioBtn.setRoundTopLeft(20);
+        imgInventarioBtn.setRoundTopRight(20);
+
+        imgClienteText1.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
+        imgClienteText1.setIcon(new javax.swing.JLabel() {
+            public javax.swing.Icon getIcon() {
+                try {
+                    return new javax.swing.ImageIcon(
+                        new java.net.URL("https://blogger.googleusercontent.com/img/a/AVvXsEgSZqCJBvZHHQYEQtjKYn3DKBsBUXujUGc7cImWwgccgKFpNmqqm0R4Iy2kOX3f0uQ2Pa6x4IFNa5zS273HvIwZ3b4TqJMbyMYHnbofbYiDA85HTNKHObpdIT1YNCtYYKVQMH86bOj8aiCl-8hw0EFJcSFqb-vfBx1XDND1V5h7t0lOyFXRumSLaA2KJPg6=w279-h193")
+                    );
+                } catch (java.net.MalformedURLException e) {
+                }
+                return null;
+            }
+        }.getIcon());
+
+        javax.swing.GroupLayout imgInventarioBtnLayout = new javax.swing.GroupLayout(imgInventarioBtn);
+        imgInventarioBtn.setLayout(imgInventarioBtnLayout);
+        imgInventarioBtnLayout.setHorizontalGroup(
+            imgInventarioBtnLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, imgInventarioBtnLayout.createSequentialGroup()
+                .addContainerGap()
+                .addComponent(imgClienteText1, javax.swing.GroupLayout.PREFERRED_SIZE, 0, Short.MAX_VALUE)
+                .addContainerGap())
+        );
+        imgInventarioBtnLayout.setVerticalGroup(
+            imgInventarioBtnLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(imgInventarioBtnLayout.createSequentialGroup()
+                .addContainerGap()
+                .addComponent(imgClienteText1, javax.swing.GroupLayout.DEFAULT_SIZE, 168, Short.MAX_VALUE))
+        );
+
+        inventarioBtn.setBackground(new java.awt.Color(255, 255, 255));
+        inventarioBtn.setRoundBottomLeft(20);
+        inventarioBtn.setRoundBottomRight(20);
+
+        inventarioText.setFont(new java.awt.Font("Segoe UI", 1, 24)); // NOI18N
+        inventarioText.setForeground(new java.awt.Color(0, 0, 0));
+        inventarioText.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
+        inventarioText.setText("Inventario");
+
+        javax.swing.GroupLayout inventarioBtnLayout = new javax.swing.GroupLayout(inventarioBtn);
+        inventarioBtn.setLayout(inventarioBtnLayout);
+        inventarioBtnLayout.setHorizontalGroup(
+            inventarioBtnLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addComponent(inventarioText, javax.swing.GroupLayout.DEFAULT_SIZE, 230, Short.MAX_VALUE)
+        );
+        inventarioBtnLayout.setVerticalGroup(
+            inventarioBtnLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, inventarioBtnLayout.createSequentialGroup()
+                .addContainerGap()
+                .addComponent(inventarioText, javax.swing.GroupLayout.DEFAULT_SIZE, 88, Short.MAX_VALUE)
+                .addContainerGap())
+        );
+
+        javax.swing.GroupLayout cardInventarioLayout = new javax.swing.GroupLayout(cardInventario);
+        cardInventario.setLayout(cardInventarioLayout);
+        cardInventarioLayout.setHorizontalGroup(
+            cardInventarioLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addComponent(imgInventarioBtn, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+            .addComponent(inventarioBtn, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+        );
+        cardInventarioLayout.setVerticalGroup(
+            cardInventarioLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(cardInventarioLayout.createSequentialGroup()
+                .addComponent(imgInventarioBtn, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addComponent(inventarioBtn, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
+        );
+
+        bg.add(cardInventario, new org.netbeans.lib.awtextra.AbsoluteConstraints(520, 270, 230, 280));
+
+        cardFactura.setBackground(new java.awt.Color(255, 255, 255));
+        cardFactura.setForeground(new java.awt.Color(255, 255, 255));
+        cardFactura.setRoundBottomLeft(20);
+        cardFactura.setRoundBottomRight(20);
+        cardFactura.setRoundTopLeft(20);
+        cardFactura.setRoundTopRight(20);
+
+        imgFacturaBtn.setBackground(new java.awt.Color(255, 255, 255));
+        imgFacturaBtn.setRoundTopLeft(20);
+        imgFacturaBtn.setRoundTopRight(20);
+
+        imgFacturaText.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
+        imgFacturaText.setIcon(new javax.swing.JLabel() {
+            public javax.swing.Icon getIcon() {
+                try {
+                    return new javax.swing.ImageIcon(
+                        new java.net.URL("https://blogger.googleusercontent.com/img/a/AVvXsEjKYrIVxmp2JzVOTRjqk8phuNdAfxhpnzel-qCEdMQslv8bwr3waEFa-qJknhNhDEQEHyU8nUmj_kDpfTtgTXsFXDiuxAGWCwYUVk7eOJKTzIA9B3uGp4K8XEPZAhFuW5Yaz2qgbfr68Mf3hhoaQApqWrwy6bACTH41R2qzp7q13dir6LIsKxdW6_dAABRR=w241-h163")
+                    );
+                } catch (java.net.MalformedURLException e) {
+                }
+                return null;
+            }
+        }.getIcon());
+
+        javax.swing.GroupLayout imgFacturaBtnLayout = new javax.swing.GroupLayout(imgFacturaBtn);
+        imgFacturaBtn.setLayout(imgFacturaBtnLayout);
+        imgFacturaBtnLayout.setHorizontalGroup(
+            imgFacturaBtnLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, imgFacturaBtnLayout.createSequentialGroup()
+                .addContainerGap()
+                .addComponent(imgFacturaText, javax.swing.GroupLayout.PREFERRED_SIZE, 0, Short.MAX_VALUE)
+                .addContainerGap())
+        );
+        imgFacturaBtnLayout.setVerticalGroup(
+            imgFacturaBtnLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(imgFacturaBtnLayout.createSequentialGroup()
+                .addContainerGap()
+                .addComponent(imgFacturaText, javax.swing.GroupLayout.DEFAULT_SIZE, 168, Short.MAX_VALUE))
+        );
+
+        facturaBtn.setBackground(new java.awt.Color(255, 255, 255));
+        facturaBtn.setRoundBottomLeft(20);
+        facturaBtn.setRoundBottomRight(20);
+
+        facturaText.setFont(new java.awt.Font("Segoe UI", 1, 24)); // NOI18N
+        facturaText.setForeground(new java.awt.Color(0, 0, 0));
+        facturaText.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
+        facturaText.setText("Facturacion");
+
+        javax.swing.GroupLayout facturaBtnLayout = new javax.swing.GroupLayout(facturaBtn);
+        facturaBtn.setLayout(facturaBtnLayout);
+        facturaBtnLayout.setHorizontalGroup(
+            facturaBtnLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addComponent(facturaText, javax.swing.GroupLayout.DEFAULT_SIZE, 230, Short.MAX_VALUE)
+        );
+        facturaBtnLayout.setVerticalGroup(
+            facturaBtnLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, facturaBtnLayout.createSequentialGroup()
+                .addContainerGap()
+                .addComponent(facturaText, javax.swing.GroupLayout.DEFAULT_SIZE, 88, Short.MAX_VALUE)
+                .addContainerGap())
+        );
+
+        javax.swing.GroupLayout cardFacturaLayout = new javax.swing.GroupLayout(cardFactura);
+        cardFactura.setLayout(cardFacturaLayout);
+        cardFacturaLayout.setHorizontalGroup(
+            cardFacturaLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addComponent(imgFacturaBtn, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+            .addComponent(facturaBtn, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+        );
+        cardFacturaLayout.setVerticalGroup(
+            cardFacturaLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(cardFacturaLayout.createSequentialGroup()
+                .addComponent(imgFacturaBtn, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addComponent(facturaBtn, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
+        );
+
+        bg.add(cardFactura, new org.netbeans.lib.awtextra.AbsoluteConstraints(940, 270, 230, 280));
+
+        javax.swing.GroupLayout layout = new javax.swing.GroupLayout(getContentPane());
+        getContentPane().setLayout(layout);
+        layout.setHorizontalGroup(
+            layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addComponent(bg, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+        );
+        layout.setVerticalGroup(
+            layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addComponent(bg, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+        );
+
+        pack();
+    }// </editor-fold>//GEN-END:initComponents
+
+    private void exitTextMouseClicked(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_exitTextMouseClicked
+        // TODO add your handling code here:
+        System.exit(0);
+
+    }//GEN-LAST:event_exitTextMouseClicked
+
+    private void exitTextMouseEntered(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_exitTextMouseEntered
+        // TODO add your handling code here:
+        exitButton.setBackground(Color.red);
+        exitText.setForeground(Color.white);
+    }//GEN-LAST:event_exitTextMouseEntered
+
+    private void exitTextMouseExited(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_exitTextMouseExited
+        // TODO add your handling code here:
+        exitButton.setBackground(Color.decode("#D95A53"));
+        exitText.setForeground(Color.white);
+    }//GEN-LAST:event_exitTextMouseExited
+
+    private void controlBarMouseDragged(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_controlBarMouseDragged
+        // TODO add your handling code here:
+        int x = evt.getXOnScreen();
+        int y = evt.getYOnScreen();
+        this.setLocation(x - xPos, y - yPos);
+    }//GEN-LAST:event_controlBarMouseDragged
+
+    private void controlBarMousePressed(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_controlBarMousePressed
+        // TODO add your handling code here:
+        xPos = evt.getX();
+        yPos = evt.getY();
+    }//GEN-LAST:event_controlBarMousePressed
+
+    public void nombreCuenta(String nombre){
+        userText.setText(nombre);
+    }
+    /**
+     * @param args the command line arguments
+     */
+    public static void main(String args[]) {
+        /* Set the Nimbus look and feel */
+        //<editor-fold defaultstate="collapsed" desc=" Look and feel setting code (optional) ">
+        /* If Nimbus (introduced in Java SE 6) is not available, stay with the default look and feel.
+         * For details see http://download.oracle.com/javase/tutorial/uiswing/lookandfeel/plaf.html 
+         */
+        try {
+            for (javax.swing.UIManager.LookAndFeelInfo info : javax.swing.UIManager.getInstalledLookAndFeels()) {
+                if ("Nimbus".equals(info.getName())) {
+                    javax.swing.UIManager.setLookAndFeel(info.getClassName());
+                    break;
+                }
+            }
+        } catch (ClassNotFoundException ex) {
+            java.util.logging.Logger.getLogger(Menu.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
+        } catch (InstantiationException ex) {
+            java.util.logging.Logger.getLogger(Menu.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
+        } catch (IllegalAccessException ex) {
+            java.util.logging.Logger.getLogger(Menu.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
+        } catch (javax.swing.UnsupportedLookAndFeelException ex) {
+            java.util.logging.Logger.getLogger(Menu.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
+        }
+        //</editor-fold>
+        //</editor-fold>
+        //</editor-fold>
+        //</editor-fold>
+        //</editor-fold>
+        //</editor-fold>
+        //</editor-fold>
+        //</editor-fold>
+        //</editor-fold>
+        //</editor-fold>
+        //</editor-fold>
+        //</editor-fold>
+        //</editor-fold>
+        //</editor-fold>
+        //</editor-fold>
+        //</editor-fold>
+
+        /* Create and display the form */
+        java.awt.EventQueue.invokeLater(new Runnable() {
+            public void run() {
+                new Menu().setVisible(true);
+            }
+        });
+    }
+
+    // Variables declaration - do not modify//GEN-BEGIN:variables
+    private javax.swing.JPanel bg;
+    private Interfaz.PanelRound cardCliente;
+    private Interfaz.PanelRound cardFactura;
+    private Interfaz.PanelRound cardInventario;
+    private Interfaz.PanelRound clienteBtn;
+    private javax.swing.JLabel clienteText;
+    private javax.swing.JPanel controlBar;
+    private javax.swing.JLabel copyRight;
+    private javax.swing.JPanel exitButton;
+    private javax.swing.JLabel exitText;
+    private Interfaz.PanelRound facturaBtn;
+    private javax.swing.JLabel facturaText;
+    private javax.swing.JPanel footer;
+    private javax.swing.JPanel header;
+    private Interfaz.PanelRound imgClienteBtn;
+    private javax.swing.JLabel imgClienteText;
+    private javax.swing.JLabel imgClienteText1;
+    private Interfaz.PanelRound imgFacturaBtn;
+    private javax.swing.JLabel imgFacturaText;
+    private Interfaz.PanelRound imgInventarioBtn;
+    private Interfaz.PanelRound inventarioBtn;
+    private javax.swing.JLabel inventarioText;
+    private javax.swing.JLabel logo;
+    private javax.swing.JLabel logoutText;
+    private javax.swing.JLabel tittle;
+    private javax.swing.JLabel userText;
+    // End of variables declaration//GEN-END:variables
+}
